requirements.txt
ollama>=0.1.2
openai~=1.68.2
python-dotenv~=1.0.1
requests~=2.32.3
torch~=2.2.2
torchvision~=0.17.2
streamlit~=1.37.1
pillow~=10.4.0
numpy~=1.26.4
pandas~=2.2.2
matplotlib~=3.9.2
openpyxl~=3.1.5
scikit-image~=0.24.0
sympy~=1.13.2
scikit-learn~=1.5.1
holoviews~=1.19.1
imageio~=2.33.1
altair~=5.0.1
json5~=0.9.6
langchain~=0.3.20
prometheus_client~=0.14.1
langchain-core~=0.3.45
langchain-openai~=0.3.8
gradio~=5.21.0
uvicorn~=0.34.0
pydantic~=2.8.2
fastapi~=0.115.11
transformers~=4.49.0
peft~=0.14.0
python-dateutil~=2.9.0post0
langchain-community~=0.3.19
pypdf>=5.4.0

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0

# 网络请求和网页处理
beautifulsoup4==4.12.3
lxml==5.1.0

# 文本处理
tiktoken

# 可选依赖（用于GPU加速）
faiss-cpu==1.8.0

# 模型访问相关
httpx==0.25.2
websockets==12.0
